LayoutTests/fast/forms/label/label-click-event-dispatch-untrusted-expected.txt

Tests that clicking a label dispatches an untrusted click event on the associated form control.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.isTrusted is false
PASS successfullyParsed is true

TEST COMPLETE
